requirements.txt
torch 
torchvision
transformers
datasets
scikit-learn
matplotlib
pillow
tqdm
seaborn